Seperated mockdata in to fixtures
