+++ tests/QuestionsAndAnswers.test.jsx
@@ -14,6 +14,10 @@ import { Provider } from 'react-redux';
 import QuestionsAndAnswers from '../client/src/components/Q&A/QuestionsAndAnswers';
 // import QuestionEntry from '../client/src/components/Q&A/QuestionEntry';
 import store from '../client/src/store/store';
+import data from './fixtures/QuestionMockData';
+
+// Destructuring data from fixtures
+const { mockQuestionData, mockAnswerData } = data;
 
 beforeAll(() => {
   axios.get.mockImplementation((url) => {
@@ -35,83 +39,6 @@ jest.mock('../client/src/components/Q&A/questions.css', () => () => (<div>Placeh
 jest.mock('../client/src/components/Q&A/QuestionEntry.jsx', () => () => (<div data-testid="question-entry">Placeholder Question Entry</div>));
 jest.mock('../client/src/components/Q&A/QuestionModal.jsx', () => () => (<div data-testid="question-modal">Placeholder Question Modal</div>));
 
-const mockQuestionData = {};
-mockQuestionData.data = {
-  product_id: 11,
-  results: [
-    {
-      question_id: 1,
-      question_body: 'Can I wash it?',
-      question_date: '2018-02-08T00:00:00.0007',
-      asker_name: 'bobby',
-      question_helpfulness: 8,
-      reported: false,
-      answers: {
-        888999: {
-          id: 888999,
-          body: 'Put it in the washer',
-          date: '2018-03-08T00:00:00.000Z',
-          answerer_name: 'jimmy',
-          helpfulness: 17,
-          photos: [],
-        },
-      },
-    },
-    {
-      question_id: 2,
-      question_body: 'Is it big or small?',
-      question_date: '2018-03-08T00:00:00.0007',
-      asker_name: 'john',
-      question_helpfulness: 12,
-      reported: false,
-      answers: {
-        123456: {
-          id: 123456,
-          body: 'It runs extra large',
-          date: '2018-04-08T00:00:00.000Z',
-          answerer_name: 'lester',
-          helpfulness: 21,
-          photos: [],
-        },
-      },
-    },
-    {
-      question_id: 3,
-      question_body: 'What material is it?',
-      question_date: '2018-05-08T00:00:00.0007',
-      asker_name: 'stephen',
-      question_helpfulness: 20,
-      reported: false,
-      answers: {
-        654321: {
-          id: 654321,
-          body: 'The finest',
-          date: '2019-04-08T00:00:00.000Z',
-          answerer_name: 'bigdan',
-          helpfulness: 2,
-          photos: [],
-        },
-      },
-    },
-  ],
-};
-
-const mockAnswerData = {
-  question: 1,
-  page: 1,
-  count: 5,
-  results: [
-    {
-      answer_id: 888999,
-      body: 'Put it in the washer',
-      date: '2018-03-08T00:00:00.000Z',
-      answerer_name: 'jimmy',
-      helpfulness: 17,
-      photos: [],
-    },
-  ],
-};
-
 // TESTS ---------------------------------------------------
 it('should load the mock questions',
   () => axios.get('/qa/questions/48432')
@@ -122,13 +49,13 @@ it('should load the mock answers',
     .then((answers) => expect(answers).toEqual(mockAnswerData)));
 
 it('should load and display the selected product data', async () => {
-  const { getByTestId } = render(
+  const { getByTestId, findAllByTestId } = render(
     <Provider store={store}>
       <QuestionsAndAnswers />
     </Provider>,
   );
   // await waitFor(() => screen.getByTestId('question-entry'));
-  const questions = await screen.findAllByTestId('question-entry');
+  const questions = await findAllByTestId('question-entry');
   // three questions as input from mock data, should only display 2 on load
   expect(questions).toHaveLength(2);
   expect(getByTestId('add-question')).toHaveTextContent('Add Question');
